// FrontBookRentals/src/app/api-client/model/bookResponseDto.ts
/**
 * Book Rentals API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BookResponseDto { 
    bookId: string;
    name: string;
    author: string;
    isbn: string;
    status: boolean;
}