// FrontBookRentals/src/app/api-client/model/clientResponseDto.ts
/**
 * Book Rentals API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ClientResponseDto { 
    clientId: string;
    names: string;
    lastNames: string;
    dni: string;
    age: number;
}